MWPW-138362 Batches to be executed based on copy and does not need to wait for preview
A state a field is added to indicate copy completion "details":{"promoteCopyStatus":"COMPLETED"}}}
Trigger and track checks for this field and if its COMPLETE would trigger the next batch while
the preview and promote is moved to postPromoteWorker.
The overall completion of batch still is tracked via batch status.
Example:
{"action":{"lastTriggeredBy":"","type":"promoteBatch_2","status":"IN PROGRESS","message":"Previewing/Publishing promoted content for 2","activationId":"","startTime":"2023-11-09T12:59:19.645Z","endTime":"","batchesInfo":[],"details":{"promoteCopyStatus":"COMPLETED"}}}